requirements.txt
dj-database-url==0.5.0
Django==2.1.5
django-filter==2.0.0
django-heroku==0.3.1
djangorestframework==3.9.1
gunicorn==19.9.0
Markdown==3.0.1
Pillow==5.4.1
psycopg2==2.7.6.1
pytz==2018.9
whitenoise==4.1.2
